It's slow as sin, it keeps nulling out the values for no reason, but it is kind of hooked into MP5, and is sometimes showing the correct data. Should probably work on performance soon...

+++ PartyDisplay/Data/IPlayer.cs
@@ -5,8 +5,8 @@ namespace PartyDisplay.Data {
         where TCharacter : ICharacter 
         where TItem : IItem {
         public string Name { get; set; }
-        public int StarCount { get; set; }
-        public int CoinCount { get; set; }
+        public short StarCount { get; set; }
+        public short CoinCount { get; set; }
         public Ranking Ranking { get; set; }
         public TCharacter Character { get; set; }
         public ObservableCollection<TItem> Items { get; set; }
