+++ PartyDisplay/Data/mp4/Mp4Player.cs
@@ -3,8 +3,8 @@
 namespace PartyDisplay.Data.mp4 {
     public class Mp4Player:IPlayer<Mp4Character, Mp4Item> {
         public string Name { get; set; }
-        public int StarCount { get; set; }
-        public int CoinCount { get; set; }
+        public short StarCount { get; set; }
+        public short CoinCount { get; set; }
         public Ranking Ranking { get; set; }
         public Mp4Character Character { get; set; }
         public ObservableCollection<Mp4Item> Items { get; set; } = [];
