+++ PartyDisplay/Data/mp6/Mp6Player.cs
@@ -3,8 +3,8 @@
 namespace PartyDisplay.Data.mp6 {
     public class Mp6Player:IPlayer<Mp6Character, Mp6Orb> {
         public string Name { get; set; }
-        public int StarCount { get; set; }
-        public int CoinCount { get; set; }
+        public short StarCount { get; set; }
+        public short CoinCount { get; set; }
         public Ranking Ranking { get; set; }
         public Mp6Character Character { get; set; }
         public ObservableCollection<Mp6Orb> Items { get; set; } = [];
