+++ PartyDisplay/Data/mp7/Mp7Player.cs
@@ -3,8 +3,8 @@
 namespace PartyDisplay.Data.mp7 {
     public class Mp7Player:IPlayer<Mp7Character, Mp7Orb> {
         public string Name { get; set; }
-        public int StarCount { get; set; }
-        public int CoinCount { get; set; }
+        public short StarCount { get; set; }
+        public short CoinCount { get; set; }
         public Ranking Ranking { get; set; }
         public Mp7Character Character { get; set; }
         public ObservableCollection<Mp7Orb> Items { get; set; } = [];
